New package: exit.
